update fragment and add SeriesDetailsFragment
