+++ app/src/main/res/layout/fragment_series_details.xml
@@ -1,0 +1,160 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+        <variable
+            name="viewModel"
+            type="com.example.marvelverse.app.ui.series.details.SeriesViewModel" />
+        <variable
+            name="item"
+            type="com.example.marvelverse.domain.entities.main.Series" />
+
+
+    </data>
+
+        <LinearLayout
+            xmlns:app="http://schemas.android.com/apk/res-auto"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:orientation="vertical"
+            tools:ignore="UselessParent">
+
+            <androidx.appcompat.widget.Toolbar
+                android:id="@+id/toolbar_details_series"
+                android:layout_width="match_parent"
+                android:layout_height="?actionBarSize"
+                app:navigationIcon="@drawable/baseline_arrow_back_24"
+                app:title="Series">
+
+
+            </androidx.appcompat.widget.Toolbar>
+
+            <androidx.core.widget.NestedScrollView
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:fillViewport="true">
+
+                <LinearLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:orientation="vertical">
+
+                    <androidx.constraintlayout.utils.widget.ImageFilterView
+                        android:id="@+id/image_series"
+                        android:layout_width="match_parent"
+                        android:layout_height="224dp"
+                        android:layout_marginHorizontal="16dp"
+                        android:layout_marginTop="8dp"
+                        android:scaleType="centerCrop"
+                        app:imageUrl="@{item.thumbnail.path+'.'+item.thumbnail.extension}"
+                        app:roundPercent="0.1"
+                        tools:src="@drawable/spider" />
+
+                    <TextView
+                        android:id="@+id/text_series_name"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginStart="16dp"
+                        android:layout_marginTop="16dp"
+                        android:fontFamily="@font/poppins_bold"
+                        android:textColor="@color/main_title"
+                        android:text="@{item.title}"
+                        android:textSize="16sp"
+                        tools:text=" Fantastic Four by Dan Slott Vol. 1 (2021)" />
+
+                    <TextView
+                        android:id="@+id/text_description"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginStart="16dp"
+                        android:layout_marginTop="16dp"
+                        android:fontFamily="@font/poppins_bold"
+                        android:textColor="@color/main_title"
+                        android:textSize="16sp"
+                        android:text="@string/description" />
+
+                    <TextView
+                        android:id="@+id/text_description_all"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginStart="16dp"
+                        android:layout_marginTop="16dp"
+                        android:fontFamily="@font/poppins_bold"
+                        android:textColor="@color/main_title"
+                        android:textSize="16sp"
+                        android:text="@{item.description}"
+                        android:alpha="0.5"
+                        tools:text="@string/descrition_all" />
+                    <TextView
+                        android:id="@+id/text_comics"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginStart="16dp"
+                        android:layout_marginTop="16dp"
+                        android:fontFamily="@font/poppins_bold"
+                        android:textColor="@color/main_title"
+                        android:textSize="16sp"
+                        android:text="@string/characters" />
+
+                    <androidx.recyclerview.widget.RecyclerView
+                        android:id="@+id/series_characters"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:layout_marginTop="8dp"
+                        android:clipToPadding="false"
+                        android:orientation="horizontal"
+                        android:paddingHorizontal="16dp"
+                        app:items="@{viewModel.character}"
+                        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+                        tools:listitem="@layout/item_character_card" />
+                    <TextView
+                        android:id="@+id/text_characters"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginStart="16dp"
+                        android:layout_marginTop="16dp"
+                        android:fontFamily="@font/poppins_bold"
+                        android:textColor="@color/main_title"
+                        android:textSize="16sp"
+                        android:text="@string/comics" />
+
+                    <androidx.recyclerview.widget.RecyclerView
+                        android:id="@+id/series_comics"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:orientation="horizontal"
+                        android:paddingHorizontal="16dp"
+                        android:clipToPadding="false"
+                        app:items="@{viewModel.comics}"
+                        tools:listitem="@layout/item_home_card"
+                        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager" />
+                    <TextView
+                        android:id="@+id/text_events"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginStart="16dp"
+                        android:layout_marginTop="16dp"
+                        android:fontFamily="@font/poppins_bold"
+                        android:textColor="@color/main_title"
+                        android:textSize="16sp"
+                        android:text="@string/events" />
+
+                    <androidx.recyclerview.widget.RecyclerView
+                        android:id="@+id/series_events"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:orientation="horizontal"
+                        tools:listitem="@layout/item_home_card"
+                        android:paddingHorizontal="16dp"
+                        android:clipToPadding="false"
+                        app:items="@{viewModel.event}"
+                        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+                        />
+
+
+
+                </LinearLayout>
+            </androidx.core.widget.NestedScrollView>
+        </LinearLayout>
+</layout>
